Fix flaky GetCmdline() tests in Github Action containers
The test for utils.GetCmdline() sometimes returns an error because the
process started by subprocess.popen() does not _yet_ return a valid
command line via /proc/$pid/cmdline. This does only seem to happen in
containers running in the Github Actions environment as I was not able
to reproduce this anywhere else. The "fix" is to wait for one second so
things can settle.

Signed-off-by: Rudolph Bott <r@bott.im>