replace simplejson with pythons buildin json
This commit replaces the simplejson library with the python build-in json module.
In the newer Python versions starting from 3.2, there is no reason to prefer simplejson to the build in
See https://bugs.python.org/issue7451

To deal with bytes, which build-in json cannot handle by default, a ByteEncoder has been implemented.
See #1427

This removes simplejson as external dependency.

Signed-off-by: codefritzel <pascal-pf@hotmail.de>