switch name- and ip-checking to opt-in (#1589)
As often discussed (at Ganeticons etc.) many people like to just create
instances without the need to resolv their name via DNS or /etc/hosts.
So a common boilerplate in the add command is to use `--no-name-check`
`--no-ip-check`. This commit inverts the logic to an opt-in model.

During changing the code, I discovered an undocumented value of the
instance's NIC: `ip=auto`, which resolves the instance name and sets the
IP to the result. Until now, I was not aware of this.

Fix the QA by blind guessing with this idea: before the name-check was
implicitly enabled. Now it must be explicitly enabled via extra CLI option.

Signed-off-by: Sascha Lucas <sascha_lucas@web.de>