kvm online disk resize via qmp
This commit adds the ability to grow a disk of a kvm instance online
without hook scripts.  When the 'gnt-instance grow-disk' command is
completed, an rpc call is called, which notifies the hypervisor of a
disk change via qmp-socket.

This was realised with the new 'abstracted' method 'ResizeDisk' in
lib/hypervisor/hv_base.py, which was currently only implemented for the
kvm hypervisor.

See #1549 & #28

Signed-off-by: codefritzel <pascal-pf@hotmail.de>