redesign kvm qmp monitor
This commit has changed the substructure of QMP communication. The code has
been modernised with typehints codestyle and better logging.

The inheritance hierarchy has been adapted so that communication with the
QGA can be implemented in the future.

Furthermore, the QmpConnection class can now wait for certain events (see
QmpConnection#wait_for_qmp_event).  This can also solve the issue #1749.

Signed-off-by: codefritzel <pascal-pf@hotmail.de>