.ghci: removed
It contained search path src/ and test/hs which interfers with `cabal repl`.
Namely, if you run `cabal repl htest` it will try to load Ganeti.BasicTypes from sources
instead of importing it from the compiled library part.
This will fail, because cabal_macros.h are set for `htest` and not the library part
and thus some package version macros are missing.