Utils.TimeDiff: simplified custom definition that avoids bug in old-time:System.Time.addToClockTime
Do the affected functions rely on time differences like "a year" or "a month"
to be sensitive to the starting date?
JQueue.archiveJobs uses addToClockTime for timeout given in seconds. Thus no problem.
XenCpuLoad.combineWithRollover uses the time difference between two time stamps.
It is likely, that it does not depend on special meaning of year or month differences.