add Haskel lib regex-tdfa to CI matrix
Recently regex-pcre is no longer available in Debian testing (Trixi) and
an alternative regex-tdfa has been implemented, but at the coast of
losing PCRE regex in favor of POSIX. Adding regex-tdfa to the CI matrix.
This means that pcre and tdfa are tested both (where possible) for a
transitional period.

Signed-off-by: Sascha Lucas <sascha_lucas@web.de>