kvm_runtime as seperate class
This commit abstracts the KVMRuntime as a separate class, which
encapsulates the structure behind it (list or dict).  Previously, the
runtime was displayed as a list or tuple. It can be assumed that this
structure is Haskell will continue to be used. For this reason, the KVM
runtime actually represents a list. In the future, this can be converted
into a dict without having to adapt the code outside KVMRuntime.

This is required for the implementation of vCPU and RAM hotplug.

Signed-off-by: codefritzel <pascal-pf@hotmail.de>