KVM: make Ganeti compatible to the q35 machine type
This commit adds preliminary compatibility for the q35 machine
type by configuring a PCIe-to-PCI bridge on the PCIe bus. NICs,
disks etc. are still located on the PCI bus and rely on the
slower SHPC mode for hotplugging (hence the increased timeout/retry
value).

While the final goal should be to use native PCIe devices / hotplugging
on the q35, this will at least enable users to use this machine type at
all with Ganeti.

Signed-off-by: Rudolph Bott <r@bott.im>