Classify DRBD devices in StandAlone mode as faulty
DRBD devices should be reported faulty not only when in diskless mode,
but also in StandAlone mode (as local data will most likely be outdated).
This fixes the issue that degraded connections will be re-classified
from ERROR to WARNING as long as they are not diskless on any of the
nodes.

Broken/disconnected DRBD devices will now generate a proper ERROR and
corresponding return code from `gnt-cluster verify`.

Fixes #1218

Signed-off-by: Rudolph Bott <r@bott.im>