[build] Refactor Haskell PCRE handling and support pcre2
Change from using multiple `--enable-haskell-pcre' flags to a single
`--with-haskell-pcre=' option and support library autodetection. This
allows ./configure to be run without specifying a pcre option and will
pick the first available supported PCRE regex library.

Note that without this, make distcheck-release on Debian fails, as it
runs ./configure without any arguments which in turn would expect the
(deprecated) haskell-pcre library to be present.

Finally, add support for pcre2 (fixes #1830, thanks to Adrian Bunk!) and
adjust the CI workflow to the new ./configure flag.

Signed-off-by: Apollon Oikonomopoulos <apoikos@dmesg.gr>