Fix ganeti-cleaner test PYTHONPATH
The ganeti-cleaner unit test expands the relative PYTHONPATH defined in
the Makefile, so that tools/check-cert-expired can find the `ganeti'
module while running in a temporary directory. This expansion was
checking for an exact PYTHONPATH match to replace and has been broken
since 9c3cbcd2f (add pytest as testing framework, 2024-07-29), resulting
in a failed test.

Fix this and make things a bit more robust against moving files around
by making the substitution a more liberal.

Signed-off-by: Apollon Oikonomopoulos <apoikos@dmesg.gr>