fix import-export CN for gnt-backup use
Commit b5ac1e7022c0244b57ca0d7bd7e21184cbbe962a fixes certificate
hostname verification for newer socat versions. But it seems that the
usecase was the `move-instance` tool for moving between separate Ganeti
clusters. However import-export is also used for `gnt-backup`, where the
internal noded key/cert is used. This cert still has CN
ganeti.example.com (as seen in `lib/backend.py`
`StartImportExportDaemon()`):

```
  if opts.key_name is None:
    # Use server.pem
    key_path = pathutils.NODED_CERT_FILE
    cert_path = pathutils.NODED_CERT_FILE
```

Take this as an indirect measurement to detect if running for
`gnt-backup` or `move-instance`.