avoid hostname resolution on remote-import
There is an extra hostname resolution for instances created in mode
remote-import (aka `move-instance`), but the reason is unknown. This
causes an error, if the instance's name can not be resolved (visible in
the job at the receiving cluster):

  Result:
    - OpPrereqError
    - - The given name (test.vm) does not resolve: Name or service not known

It seems to work without hostname resolution, so removing it here.

Signed-off-by: Sascha Lucas <sascha_lucas@web.de>