hscan: save corrected fixdata (#1595)
When PR #34 was merged back in 2017, `hscan` was updated to look at the
new data and present it to stdout, however the change was never applied
to the saved copy (LOCAL.data) of the information.

When this incorrect data is reloaded through hbal, under some conditions
the invalid memory information can cause it to silently stop considering
otherwise viable instance moves.

Correcting hscan to actually save the fixed data for external use,
instead of saving the unmodified copy keeps things working as expected.

Co-authored-by: Ken Sanislo <kens@flyingcroc.net>