mark qmp connection for close
commit 1d8adf0b70b2a56f42cbff25d83c2be83a43da95 has redesigned the KVM
QMP monitor. It seems, that it was forgotten to mark the connection for
closing at the end. This yields to blocking QMP socket, i.e. at instance
shutdown, where frequent QMP connections sent the power down command
competing with other QMP boilerplate (query balloon and cpu fast):

A running instance with no OS inside will timeout regularly on
shutdown, but often the default timeout of 120s is not reached:

        root@master:~# gnt-instance shutdown test.vm
        Waiting for job 51 for test.vm ...
        Job 51 for test.vm has failed: Failure: command execution error:
        Could not shutdown instance 'test.vm': Error while executing backend function: [Errno 11] Resource temporarily unavailable

Looking at node-log in debug mode, one can see, that the close function
is not called (no `Socket Connection to %s closed` in the log). With
this fix, the connection is closed and shutdown / reboot works as
expected.

Signed-off-by: Sascha Lucas <sascha_lucas@web.de>