Upgrade KVM runtime WRT to removal of `-chroot` (#1853)
Commit e607423b57521fec5a9a1d289ca736ec90266b18 has removed `-chroot`
from the KVM runtime, but did not take into account, that instances
started by <Qemu-9.0 needs to upgrade their runtime, too. Without that
fix, a rolling upgrade to >=Qemu-9.0 wont be possible, if instances are
using chroot.

Simple tests proved that chroot can be changed during live migration and
doesn't need to be in the KVM runtime. However, no further knowledge
exists why `-chroot` was in the runtime before.

Signed-off-by: Sascha Lucas <sascha_lucas@web.de>